refactor : chatrooms 페이지에서 chatrooms contents 컴포넌트 분리 및 적용
