+++ src/components/ChatroomsContent/ChatroomsContent.style.tsx
@@ -1,0 +1,11 @@
+import styled, { css } from 'styled-components';
+
+export const ContentsWrapper = styled.div`
+  ${({ theme: { defaultPadding } }) => css`
+    ${defaultPadding};
+  `}
+`;
+
+export const AdditionalModalStyle = css`
+  padding: 0 !important;
+`;
